requirements.txt
alembic==1.10.4
anyio==3.6.2
asyncpg==0.27.0
click==8.1.3
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.1
passlib==1.7.4
pluggy==1.0.0
pyasn1==0.5.0
pydantic==1.10.7
PyJWT==2.6.0
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-env==0.8.1
python-dotenv==0.15.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.11
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.21.1
